requirements.txt
requests
tqdm
transformers
pytesseract
opencv-python
matplotlib
Pillow
numpy
